+++ endpoints/openrtb2/auction.go
@@ -20,9 +20,6 @@ import (
 	"github.com/julienschmidt/httprouter"
 	gpplib "github.com/prebid/go-gpp"
 	"github.com/prebid/go-gpp/constants"
-	"github.com/prebid/openrtb/v20/adcom1"
-	"github.com/prebid/openrtb/v20/native1"
-	nativeRequests "github.com/prebid/openrtb/v20/native1/request"
 	"github.com/prebid/openrtb/v20/openrtb2"
 	"github.com/prebid/openrtb/v20/openrtb3"
 	"github.com/prebid/prebid-server/v2/bidadjustment"
@@ -923,7 +920,8 @@ func (deps *endpointDeps) validateRequest(account *config.Account, httpReq *http
 		}
 		impIDs[imp.ID] = i
 
-		errs := deps.validateImp(imp, requestAliases, i, hasStoredResponses, storedBidResp)
+		reqValidator := ortb.NewRequestValidator(deps.bidderMap, deps.disabledBidders, deps.paramsValidator)
+		errs := reqValidator.ValidateImp(imp, i, requestAliases, hasStoredResponses, storedBidResp)
 		if len(errs) > 0 {
 			errL = append(errL, errs...)
 		}
@@ -1066,536 +1064,6 @@ func (deps *endpointDeps) validateBidders(bidders []string, knownBidders map[str
 		}
 	}
 	return nil
-}
-
-func (deps *endpointDeps) validateImp(imp *openrtb_ext.ImpWrapper, aliases map[string]string, index int, hasStoredResponses bool, storedBidResp stored_responses.ImpBidderStoredResp) []error {
-	if imp.ID == "" {
-		return []error{fmt.Errorf("request.imp[%d] missing required field: \"id\"", index)}
-	}
-
-	if len(imp.Metric) != 0 {
-		return []error{fmt.Errorf("request.imp[%d].metric is not yet supported by prebid-server. Support may be added in the future", index)}
-	}
-
-	if imp.Banner == nil && imp.Video == nil && imp.Audio == nil && imp.Native == nil {
-		return []error{fmt.Errorf("request.imp[%d] must contain at least one of \"banner\", \"video\", \"audio\", or \"native\"", index)}
-	}
-
-	if err := validateBanner(imp.Banner, index, isInterstitial(imp)); err != nil {
-		return []error{err}
-	}
-
-	if err := validateVideo(imp.Video, index); err != nil {
-		return []error{err}
-	}
-
-	if err := validateAudio(imp.Audio, index); err != nil {
-		return []error{err}
-	}
-
-	if err := fillAndValidateNative(imp.Native, index); err != nil {
-		return []error{err}
-	}
-
-	if err := validatePmp(imp.PMP, index); err != nil {
-		return []error{err}
-	}
-
-	errL := deps.validateImpExt(imp, aliases, index, hasStoredResponses, storedBidResp)
-	if len(errL) != 0 {
-		return errL
-	}
-
-	return nil
-}
-
-func isInterstitial(imp *openrtb_ext.ImpWrapper) bool {
-	return imp.Instl == 1
-}
-
-func validateBanner(banner *openrtb2.Banner, impIndex int, isInterstitial bool) error {
-	if banner == nil {
-		return nil
-	}
-
-	// The following fields were previously uints in the OpenRTB library we use, but have
-	// since been changed to ints. We decided to maintain the non-negative check.
-	if banner.W != nil && *banner.W < 0 {
-		return fmt.Errorf("request.imp[%d].banner.w must be a positive number", impIndex)
-	}
-	if banner.H != nil && *banner.H < 0 {
-		return fmt.Errorf("request.imp[%d].banner.h must be a positive number", impIndex)
-	}
-
-	// The following fields are deprecated in the OpenRTB 2.5 spec but are still present
-	// in the OpenRTB library we use. Enforce they are not specified.
-	if banner.WMin != 0 {
-		return fmt.Errorf("request.imp[%d].banner uses unsupported property: \"wmin\". Use the \"format\" array instead.", impIndex)
-	}
-	if banner.WMax != 0 {
-		return fmt.Errorf("request.imp[%d].banner uses unsupported property: \"wmax\". Use the \"format\" array instead.", impIndex)
-	}
-	if banner.HMin != 0 {
-		return fmt.Errorf("request.imp[%d].banner uses unsupported property: \"hmin\". Use the \"format\" array instead.", impIndex)
-	}
-	if banner.HMax != 0 {
-		return fmt.Errorf("request.imp[%d].banner uses unsupported property: \"hmax\". Use the \"format\" array instead.", impIndex)
-	}
-
-	hasRootSize := banner.H != nil && banner.W != nil && *banner.H > 0 && *banner.W > 0
-	if !hasRootSize && len(banner.Format) == 0 && !isInterstitial {
-		return fmt.Errorf("request.imp[%d].banner has no sizes. Define \"w\" and \"h\", or include \"format\" elements.", impIndex)
-	}
-
-	for i, format := range banner.Format {
-		if err := validateFormat(&format, impIndex, i); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func validateVideo(video *openrtb2.Video, impIndex int) error {
-	if video == nil {
-		return nil
-	}
-
-	if len(video.MIMEs) < 1 {
-		return fmt.Errorf("request.imp[%d].video.mimes must contain at least one supported MIME type", impIndex)
-	}
-
-	// The following fields were previously uints in the OpenRTB library we use, but have
-	// since been changed to ints. We decided to maintain the non-negative check.
-	if video.W != nil && *video.W < 0 {
-		return fmt.Errorf("request.imp[%d].video.w must be a positive number", impIndex)
-	}
-	if video.H != nil && *video.H < 0 {
-		return fmt.Errorf("request.imp[%d].video.h must be a positive number", impIndex)
-	}
-	if video.MinBitRate < 0 {
-		return fmt.Errorf("request.imp[%d].video.minbitrate must be a positive number", impIndex)
-	}
-	if video.MaxBitRate < 0 {
-		return fmt.Errorf("request.imp[%d].video.maxbitrate must be a positive number", impIndex)
-	}
-
-	return nil
-}
-
-func validateAudio(audio *openrtb2.Audio, impIndex int) error {
-	if audio == nil {
-		return nil
-	}
-
-	if len(audio.MIMEs) < 1 {
-		return fmt.Errorf("request.imp[%d].audio.mimes must contain at least one supported MIME type", impIndex)
-	}
-
-	// The following fields were previously uints in the OpenRTB library we use, but have
-	// since been changed to ints. We decided to maintain the non-negative check.
-	if audio.Sequence < 0 {
-		return fmt.Errorf("request.imp[%d].audio.sequence must be a positive number", impIndex)
-	}
-	if audio.MaxSeq < 0 {
-		return fmt.Errorf("request.imp[%d].audio.maxseq must be a positive number", impIndex)
-	}
-	if audio.MinBitrate < 0 {
-		return fmt.Errorf("request.imp[%d].audio.minbitrate must be a positive number", impIndex)
-	}
-	if audio.MaxBitrate < 0 {
-		return fmt.Errorf("request.imp[%d].audio.maxbitrate must be a positive number", impIndex)
-	}
-
-	return nil
-}
-
-// fillAndValidateNative validates the request, and assigns the Asset IDs as recommended by the Native v1.2 spec.
-func fillAndValidateNative(n *openrtb2.Native, impIndex int) error {
-	if n == nil {
-		return nil
-	}
-
-	if len(n.Request) == 0 {
-		return fmt.Errorf("request.imp[%d].native missing required property \"request\"", impIndex)
-	}
-	var nativePayload nativeRequests.Request
-	if err := jsonutil.UnmarshalValid(json.RawMessage(n.Request), &nativePayload); err != nil {
-		return err
-	}
-
-	if err := validateNativeContextTypes(nativePayload.Context, nativePayload.ContextSubType, impIndex); err != nil {
-		return err
-	}
-	if err := validateNativePlacementType(nativePayload.PlcmtType, impIndex); err != nil {
-		return err
-	}
-	if err := fillAndValidateNativeAssets(nativePayload.Assets, impIndex); err != nil {
-		return err
-	}
-	if err := validateNativeEventTrackers(nativePayload.EventTrackers, impIndex); err != nil {
-		return err
-	}
-
-	serialized, err := jsonutil.Marshal(nativePayload)
-	if err != nil {
-		return err
-	}
-	n.Request = string(serialized)
-	return nil
-}
-
-func validateNativeContextTypes(cType native1.ContextType, cSubtype native1.ContextSubType, impIndex int) error {
-	if cType == 0 {
-		// Context is only recommended, so none is a valid type.
-		return nil
-	}
-	if cType < native1.ContextTypeContent || (cType > native1.ContextTypeProduct && cType < openrtb_ext.NativeExchangeSpecificLowerBound) {
-		return fmt.Errorf("request.imp[%d].native.request.context is invalid. See https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=39", impIndex)
-	}
-	if cSubtype < 0 {
-		return fmt.Errorf("request.imp[%d].native.request.contextsubtype value can't be less than 0. See https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=39", impIndex)
-	}
-	if cSubtype == 0 {
-		return nil
-	}
-	if cSubtype >= native1.ContextSubTypeGeneral && cSubtype <= native1.ContextSubTypeUserGenerated {
-		if cType != native1.ContextTypeContent && cType < openrtb_ext.NativeExchangeSpecificLowerBound {
-			return fmt.Errorf("request.imp[%d].native.request.context is %d, but contextsubtype is %d. This is an invalid combination. See https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=39", impIndex, cType, cSubtype)
-		}
-		return nil
-	}
-	if cSubtype >= native1.ContextSubTypeSocial && cSubtype <= native1.ContextSubTypeChat {
-		if cType != native1.ContextTypeSocial && cType < openrtb_ext.NativeExchangeSpecificLowerBound {
-			return fmt.Errorf("request.imp[%d].native.request.context is %d, but contextsubtype is %d. This is an invalid combination. See https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=39", impIndex, cType, cSubtype)
-		}
-		return nil
-	}
-	if cSubtype >= native1.ContextSubTypeSelling && cSubtype <= native1.ContextSubTypeProductReview {
-		if cType != native1.ContextTypeProduct && cType < openrtb_ext.NativeExchangeSpecificLowerBound {
-			return fmt.Errorf("request.imp[%d].native.request.context is %d, but contextsubtype is %d. This is an invalid combination. See https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=39", impIndex, cType, cSubtype)
-		}
-		return nil
-	}
-	if cSubtype >= openrtb_ext.NativeExchangeSpecificLowerBound {
-		return nil
-	}
-
-	return fmt.Errorf("request.imp[%d].native.request.contextsubtype is invalid. See https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=39", impIndex)
-}
-
-func validateNativePlacementType(pt native1.PlacementType, impIndex int) error {
-	if pt == 0 {
-		// Placement Type is only recommended, not required.
-		return nil
-	}
-	if pt < native1.PlacementTypeFeed || (pt > native1.PlacementTypeRecommendationWidget && pt < openrtb_ext.NativeExchangeSpecificLowerBound) {
-		return fmt.Errorf("request.imp[%d].native.request.plcmttype is invalid. See https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=40", impIndex)
-	}
-	return nil
-}
-
-func fillAndValidateNativeAssets(assets []nativeRequests.Asset, impIndex int) error {
-	if len(assets) < 1 {
-		return fmt.Errorf("request.imp[%d].native.request.assets must be an array containing at least one object", impIndex)
-	}
-
-	assetIDs := make(map[int64]struct{}, len(assets))
-
-	// If none of the asset IDs are defined by the caller, then prebid server should assign its own unique IDs. But
-	// if the caller did assign its own asset IDs, then prebid server will respect those IDs
-	assignAssetIDs := true
-	for i := 0; i < len(assets); i++ {
-		assignAssetIDs = assignAssetIDs && (assets[i].ID == 0)
-	}
-
-	for i := 0; i < len(assets); i++ {
-		if err := validateNativeAsset(assets[i], impIndex, i); err != nil {
-			return err
-		}
-
-		if assignAssetIDs {
-			assets[i].ID = int64(i)
-			continue
-		}
-
-		// Each asset should have a unique ID thats assigned by the caller
-		if _, ok := assetIDs[assets[i].ID]; ok {
-			return fmt.Errorf("request.imp[%d].native.request.assets[%d].id is already being used by another asset. Each asset ID must be unique.", impIndex, i)
-		}
-
-		assetIDs[assets[i].ID] = struct{}{}
-	}
-
-	return nil
-}
-
-func validateNativeAsset(asset nativeRequests.Asset, impIndex int, assetIndex int) error {
-	assetErr := "request.imp[%d].native.request.assets[%d] must define exactly one of {title, img, video, data}"
-	foundType := false
-
-	if asset.Title != nil {
-		foundType = true
-		if err := validateNativeAssetTitle(asset.Title, impIndex, assetIndex); err != nil {
-			return err
-		}
-	}
-
-	if asset.Img != nil {
-		if foundType {
-			return fmt.Errorf(assetErr, impIndex, assetIndex)
-		}
-		foundType = true
-		if err := validateNativeAssetImage(asset.Img, impIndex, assetIndex); err != nil {
-			return err
-		}
-	}
-
-	if asset.Video != nil {
-		if foundType {
-			return fmt.Errorf(assetErr, impIndex, assetIndex)
-		}
-		foundType = true
-		if err := validateNativeAssetVideo(asset.Video, impIndex, assetIndex); err != nil {
-			return err
-		}
-	}
-
-	if asset.Data != nil {
-		if foundType {
-			return fmt.Errorf(assetErr, impIndex, assetIndex)
-		}
-		foundType = true
-		if err := validateNativeAssetData(asset.Data, impIndex, assetIndex); err != nil {
-			return err
-		}
-	}
-
-	if !foundType {
-		return fmt.Errorf(assetErr, impIndex, assetIndex)
-	}
-
-	return nil
-}
-
-func validateNativeEventTrackers(trackers []nativeRequests.EventTracker, impIndex int) error {
-	for i := 0; i < len(trackers); i++ {
-		if err := validateNativeEventTracker(trackers[i], impIndex, i); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func validateNativeAssetTitle(title *nativeRequests.Title, impIndex int, assetIndex int) error {
-	if title.Len < 1 {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].title.len must be a positive number", impIndex, assetIndex)
-	}
-	return nil
-}
-
-func validateNativeEventTracker(tracker nativeRequests.EventTracker, impIndex int, eventIndex int) error {
-	if tracker.Event < native1.EventTypeImpression || (tracker.Event > native1.EventTypeViewableVideo50 && tracker.Event < openrtb_ext.NativeExchangeSpecificLowerBound) {
-		return fmt.Errorf("request.imp[%d].native.request.eventtrackers[%d].event is invalid. See section 7.6: https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=43", impIndex, eventIndex)
-	}
-	if len(tracker.Methods) < 1 {
-		return fmt.Errorf("request.imp[%d].native.request.eventtrackers[%d].method is required. See section 7.7: https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=43", impIndex, eventIndex)
-	}
-	for methodIndex, method := range tracker.Methods {
-		if method < native1.EventTrackingMethodImage || (method > native1.EventTrackingMethodJS && method < openrtb_ext.NativeExchangeSpecificLowerBound) {
-			return fmt.Errorf("request.imp[%d].native.request.eventtrackers[%d].methods[%d] is invalid. See section 7.7: https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=43", impIndex, eventIndex, methodIndex)
-		}
-	}
-
-	return nil
-}
-
-func validateNativeAssetImage(img *nativeRequests.Image, impIndex int, assetIndex int) error {
-	if img.W < 0 {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].img.w must be a positive integer", impIndex, assetIndex)
-	}
-	if img.H < 0 {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].img.h must be a positive integer", impIndex, assetIndex)
-	}
-	if img.WMin < 0 {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].img.wmin must be a positive integer", impIndex, assetIndex)
-	}
-	if img.HMin < 0 {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].img.hmin must be a positive integer", impIndex, assetIndex)
-	}
-	return nil
-}
-
-func validateNativeAssetVideo(video *nativeRequests.Video, impIndex int, assetIndex int) error {
-	if len(video.MIMEs) < 1 {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].video.mimes must be an array with at least one MIME type", impIndex, assetIndex)
-	}
-	if video.MinDuration < 1 {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].video.minduration must be a positive integer", impIndex, assetIndex)
-	}
-	if video.MaxDuration < 1 {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].video.maxduration must be a positive integer", impIndex, assetIndex)
-	}
-	if err := validateNativeVideoProtocols(video.Protocols, impIndex, assetIndex); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func validateNativeAssetData(data *nativeRequests.Data, impIndex int, assetIndex int) error {
-	if data.Type < native1.DataAssetTypeSponsored || (data.Type > native1.DataAssetTypeCTAText && data.Type < 500) {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].data.type is invalid. See section 7.4: https://iabtechlab.com/wp-content/uploads/2016/07/OpenRTB-Native-Ads-Specification-Final-1.2.pdf#page=40", impIndex, assetIndex)
-	}
-
-	return nil
-}
-
-func validateNativeVideoProtocols(protocols []adcom1.MediaCreativeSubtype, impIndex int, assetIndex int) error {
-	if len(protocols) < 1 {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].video.protocols must be an array with at least one element", impIndex, assetIndex)
-	}
-	for i := 0; i < len(protocols); i++ {
-		if err := validateNativeVideoProtocol(protocols[i], impIndex, assetIndex, i); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func validateNativeVideoProtocol(protocol adcom1.MediaCreativeSubtype, impIndex int, assetIndex int, protocolIndex int) error {
-	if protocol < adcom1.CreativeVAST10 || protocol > adcom1.CreativeDAAST10Wrapper {
-		return fmt.Errorf("request.imp[%d].native.request.assets[%d].video.protocols[%d] is invalid. See Section 5.8: https://www.iab.com/wp-content/uploads/2016/03/OpenRTB-API-Specification-Version-2-5-FINAL.pdf#page=52", impIndex, assetIndex, protocolIndex)
-	}
-	return nil
-}
-
-func validateFormat(format *openrtb2.Format, impIndex, formatIndex int) error {
-	usesHW := format.W != 0 || format.H != 0
-	usesRatios := format.WMin != 0 || format.WRatio != 0 || format.HRatio != 0
-
-	// The following fields were previously uints in the OpenRTB library we use, but have
-	// since been changed to ints. We decided to maintain the non-negative check.
-	if format.W < 0 {
-		return fmt.Errorf("request.imp[%d].banner.format[%d].w must be a positive number", impIndex, formatIndex)
-	}
-	if format.H < 0 {
-		return fmt.Errorf("request.imp[%d].banner.format[%d].h must be a positive number", impIndex, formatIndex)
-	}
-	if format.WRatio < 0 {
-		return fmt.Errorf("request.imp[%d].banner.format[%d].wratio must be a positive number", impIndex, formatIndex)
-	}
-	if format.HRatio < 0 {
-		return fmt.Errorf("request.imp[%d].banner.format[%d].hratio must be a positive number", impIndex, formatIndex)
-	}
-	if format.WMin < 0 {
-		return fmt.Errorf("request.imp[%d].banner.format[%d].wmin must be a positive number", impIndex, formatIndex)
-	}
-
-	if usesHW && usesRatios {
-		return fmt.Errorf("Request imp[%d].banner.format[%d] should define *either* {w, h} *or* {wmin, wratio, hratio}, but not both. If both are valid, send two \"format\" objects in the request.", impIndex, formatIndex)
-	}
-	if !usesHW && !usesRatios {
-		return fmt.Errorf("Request imp[%d].banner.format[%d] should define *either* {w, h} (for static size requirements) *or* {wmin, wratio, hratio} (for flexible sizes) to be non-zero.", impIndex, formatIndex)
-	}
-	if usesHW && (format.W == 0 || format.H == 0) {
-		return fmt.Errorf("Request imp[%d].banner.format[%d] must define non-zero \"h\" and \"w\" properties.", impIndex, formatIndex)
-	}
-	if usesRatios && (format.WMin == 0 || format.WRatio == 0 || format.HRatio == 0) {
-		return fmt.Errorf("Request imp[%d].banner.format[%d] must define non-zero \"wmin\", \"wratio\", and \"hratio\" properties.", impIndex, formatIndex)
-	}
-	return nil
-}
-
-func validatePmp(pmp *openrtb2.PMP, impIndex int) error {
-	if pmp == nil {
-		return nil
-	}
-
-	for dealIndex, deal := range pmp.Deals {
-		if deal.ID == "" {
-			return fmt.Errorf("request.imp[%d].pmp.deals[%d] missing required field: \"id\"", impIndex, dealIndex)
-		}
-	}
-	return nil
-}
-
-func (deps *endpointDeps) validateImpExt(imp *openrtb_ext.ImpWrapper, aliases map[string]string, impIndex int, hasStoredResponses bool, storedBidResp stored_responses.ImpBidderStoredResp) []error {
-	if len(imp.Ext) == 0 {
-		return []error{fmt.Errorf("request.imp[%d].ext is required", impIndex)}
-	}
-
-	impExt, err := imp.GetImpExt()
-	if err != nil {
-		return []error{err}
-	}
-
-	prebid := impExt.GetOrCreatePrebid()
-	prebidModified := false
-
-	if prebid.Bidder == nil {
-		prebid.Bidder = make(map[string]json.RawMessage)
-	}
-
-	ext := impExt.GetExt()
-	extModified := false
-
-	// promote imp[].ext.BIDDER to newer imp[].ext.prebid.bidder.BIDDER location, with the later taking precedence
-	for k, v := range ext {
-		if isPossibleBidder(k) {
-			if _, exists := prebid.Bidder[k]; !exists {
-				prebid.Bidder[k] = v
-				prebidModified = true
-			}
-			delete(ext, k)
-			extModified = true
-		}
-	}
-
-	if hasStoredResponses && prebid.StoredAuctionResponse == nil {
-		return []error{fmt.Errorf("request validation failed. The StoredAuctionResponse.ID field must be completely present with, or completely absent from, all impressions in request. No StoredAuctionResponse data found for request.imp[%d].ext.prebid \n", impIndex)}
-	}
-
-	if err := deps.validateStoredBidRespAndImpExtBidders(prebid, storedBidResp, imp.ID); err != nil {
-		return []error{err}
-	}
-
-	errL := []error{}
-
-	for bidder, ext := range prebid.Bidder {
-		coreBidder, _ := openrtb_ext.NormalizeBidderName(bidder)
-		if tmp, isAlias := aliases[bidder]; isAlias {
-			coreBidder = openrtb_ext.BidderName(tmp)
-		}
-
-		if coreBidderNormalized, isValid := deps.bidderMap[coreBidder.String()]; isValid {
-			if err := deps.paramsValidator.Validate(coreBidderNormalized, ext); err != nil {
-				return []error{fmt.Errorf("request.imp[%d].ext.prebid.bidder.%s failed validation.\n%v", impIndex, bidder, err)}
-			}
-		} else {
-			if msg, isDisabled := deps.disabledBidders[bidder]; isDisabled {
-				errL = append(errL, &errortypes.BidderTemporarilyDisabled{Message: msg})
-				delete(prebid.Bidder, bidder)
-				prebidModified = true
-			} else {
-				return []error{fmt.Errorf("request.imp[%d].ext.prebid.bidder contains unknown bidder: %s. Did you forget an alias in request.ext.prebid.aliases?", impIndex, bidder)}
-			}
-		}
-	}
-
-	if len(prebid.Bidder) == 0 {
-		errL = append(errL, fmt.Errorf("request.imp[%d].ext.prebid.bidder must contain at least one bidder", impIndex))
-		return errL
-	}
-
-	if prebidModified {
-		impExt.SetPrebid(prebid)
-	}
-	if extModified {
-		impExt.SetExt(ext)
-	}
-
-	return errL
 }
 
 // isPossibleBidder determines if a bidder name is a potential real bidder.
@@ -2567,33 +2035,4 @@ func checkIfAppRequest(request []byte) (bool, error) {
 		return true, nil
 	}
 	return false, nil
-}
-
-func (deps *endpointDeps) validateStoredBidRespAndImpExtBidders(prebid *openrtb_ext.ExtImpPrebid, storedBidResp stored_responses.ImpBidderStoredResp, impId string) error {
-	if storedBidResp == nil && len(prebid.StoredBidResponse) == 0 {
-		return nil
-	}
-
-	if storedBidResp == nil {
-		return generateStoredBidResponseValidationError(impId)
-	}
-	if bidResponses, ok := storedBidResp[impId]; ok {
-		if len(bidResponses) != len(prebid.Bidder) {
-			return generateStoredBidResponseValidationError(impId)
-		}
-
-		for bidderName := range bidResponses {
-			if _, bidderNameOk := deps.normalizeBidderName(bidderName); !bidderNameOk {
-				return fmt.Errorf(`unrecognized bidder "%v"`, bidderName)
-			}
-			if _, present := prebid.Bidder[bidderName]; !present {
-				return generateStoredBidResponseValidationError(impId)
-			}
-		}
-	}
-	return nil
-}
-
-func generateStoredBidResponseValidationError(impID string) error {
-	return fmt.Errorf("request validation failed. Stored bid responses are specified for imp %s. Bidders specified in imp.ext should match with bidders specified in imp.ext.prebid.storedbidresponse", impID)
 }
